+++ mutations/ag_places_off_by_one.pl
@@ -1,0 +1,2 @@
+# AG: let one more approval through than there are places.
+s{return self::places_taken\(\$instance\) >= \$places;}{return self::places_taken(\$instance) > \$places;};